feat: encode Decimal::NaN larger than +Infinity (#3)
By default, PostgreSQL treats +Infinity < NaN < Null, while SQLite
treats Null < NaN < -Infinity.

This crate does not handle `null`. For `f32` and `f64`, it treats `NaN`
as larger than `Infinity`; but for `decimal`, it treats `NaN` as smaller
than `-Infinity`.

This PR changes the ordering behavior of `decimal` to also treat `NaN`
as larger than `Infinity`.

Note that IEEE 754 does define a [total
ordering](https://doc.rust-lang.org/std/primitive.f32.html#method.total_cmp),
where `NaN` with a negative sign is smaller than `-Infinity`, and `NaN`
without sign bit set is larger than `+Infinity`. It also defines `-0` to
be smaller than `0`. This crate is already making the encoding of all
`NaN`s to be the same, and `-0` = `0`. It is not a goal to follow that
total ordering definition. (Or we could impl the standard here, and
require the caller side to handle these values before encoding.)